moved some components one level up
